# Improve performance of SimpleTriggerImpl