Improve performance of **SimpleTriggerImpl** by:
* avoiding to repeatedly invoke virtual properties.
* using `Nullable<T>.GetValueOrDefault()` when we already verified that we have a value.
* simplifying conditions of while loops, and avoid entering while loops if possible.
* avoiding unnecessary calls to `SystemTime.UtcNow`.
* avoiding to compute the number of times fired for **RescheduleNowWithRemainingRepeatCount** when **RepeatCount** is **zero** or **RepeatIndefinitely**.

The improvement is bigger on .NET Framework than on .NET (Core).

**Important:**
The **endUtc** argument of `GetFireTimeBefore(...)` is no longer nullable. When we invoke this method, we always have a value for this argument. Let me know if you want me to revert that change.

<details>
<summary>Benchmark results</summary>

**.NET 5.0**
|                                                                                                                                          Method | Branch |          Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------------------------------------------------------------------------------------------------------ |--------|--------------:|---------:|---------:|------:|------:|------:|----------:|
|                                     GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsNull | PR     |      49.50 ns | 0.096 ns | 0.080 ns |     - |     - |     - |         - |
|                                     GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsNull | master |      56.94 ns | 0.064 ns | 0.054 ns |     - |     - |     - |         - |
|    GetFireTimeAfter_RepeatCountIsGreaterThanZeroAndLessThanNumberOfTimesExecuted_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | PR     |      28.80 ns | 0.041 ns | 0.037 ns |     - |     - |     - |         - |
|    GetFireTimeAfter_RepeatCountIsGreaterThanZeroAndLessThanNumberOfTimesExecuted_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | master |      34.71 ns | 0.048 ns | 0.043 ns |     - |     - |     - |         - |
|                                                            GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsNull | PR     |     118.33 ns | 0.205 ns | 0.160 ns |     - |     - |     - |         - |
|                                                            GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsNull | master |     126.48 ns | 0.191 ns | 0.179 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | PR     |      56.49 ns | 0.229 ns | 0.214 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | master |      61.70 ns | 0.176 ns | 0.156 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMinValue | PR     |      19.47 ns | 0.026 ns | 0.025 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMinValue | master |      21.33 ns | 0.014 ns | 0.012 ns |     - |     - |     - |         - |
|                                                                    GetFireTimeAfter_RepeatCountIsZero_AfterTimeUtcIsNotNullAndAfterStartTimeUtc | PR     |      14.90 ns | 0.009 ns | 0.008 ns |     - |     - |     - |         - |
|                                                                    GetFireTimeAfter_RepeatCountIsZero_AfterTimeUtcIsNotNullAndAfterStartTimeUtc | master |      14.49 ns | 0.020 ns | 0.019 ns |     - |     - |     - |         - |
|                                                        GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsMaxValue | PR     |     124.37 ns | 0.151 ns | 0.134 ns |     - |     - |     - |         - |
|                                                        GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsMaxValue | master |     131.61 ns | 1.113 ns | 1.041 ns |     - |     - |     - |         - |
|                                    GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndBeforeStartTimeUtc_EndTimeUtcIsNull | PR     |      20.37 ns | 0.024 ns | 0.020 ns |     - |     - |     - |         - |
|                                    GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndBeforeStartTimeUtc_EndTimeUtcIsNull | master |      23.31 ns | 0.033 ns | 0.029 ns |     - |     - |     - |         - |
|                                                                       Triggered_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsNull_CalendarIsNull | PR     |      47.22 ns | 0.051 ns | 0.048 ns |     - |     - |     - |         - |
|                                                                       Triggered_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsNull_CalendarIsNull | master |      53.94 ns | 0.155 ns | 0.138 ns |     - |     - |     - |         - |
|                                                                                                UpdateAfterMisfire_SmartPolicy_RepeatCountIsZero | PR     |      81.31 ns | 0.131 ns | 0.122 ns |     - |     - |     - |         - |
|                                                                                                UpdateAfterMisfire_SmartPolicy_RepeatCountIsZero | master |      82.47 ns | 0.261 ns | 0.231 ns |     - |     - |     - |         - |
|                                                                                  UpdateAfterMisfire_SmartPolicy_RepeatCountIsRepeatIndefinitely | PR     |     142.42 ns | 0.296 ns | 0.262 ns |     - |     - |     - |         - |
|                                                                                  UpdateAfterMisfire_SmartPolicy_RepeatCountIsRepeatIndefinitely | master |     155.28 ns | 0.345 ns | 0.323 ns |     - |     - |     - |         - |
|                                                                                     UpdateAfterMisfire_SmartPolicy_RepeatCountIsGreaterThanZero | PR     |     109.43 ns | 0.133 ns | 0.118 ns |     - |     - |     - |         - |
|                                                                                     UpdateAfterMisfire_SmartPolicy_RepeatCountIsGreaterThanZero | master |     109.79 ns | 0.202 ns | 0.189 ns |     - |     - |     - |         - |
|                                                                 UpdateAfterMisfire_FireNow_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMinValue | PR     |      94.02 ns | 0.159 ns | 0.149 ns |     - |     - |     - |         - |
|                                                                 UpdateAfterMisfire_FireNow_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMinValue | master |     112.09 ns | 0.116 ns | 0.097 ns |     - |     - |     - |         - |
|                                                                           UpdateAfterMisfire_FireNow_RepeatCountIsMaxValue_EndTimeUtcIsMaxValue | PR     |     125.59 ns | 0.203 ns | 0.190 ns |     - |     - |     - |         - |
|                                                                           UpdateAfterMisfire_FireNow_RepeatCountIsMaxValue_EndTimeUtcIsMaxValue | master |     126.56 ns | 0.093 ns | 0.083 ns |     - |     - |     - |         - |
|                                         UpdateAfterMisfire_RescheduleNextWithExistingCount_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMaxValue | PR     |     129.45 ns | 0.180 ns | 0.160 ns |     - |     - |     - |         - |
|                                         UpdateAfterMisfire_RescheduleNextWithExistingCount_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMaxValue | master |     141.34 ns | 0.493 ns | 0.437 ns |     - |     - |     - |         - |

**.NET Framework 4.8**

|                                                                                                                                          Method | Branch |          Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------------------------------------------------------------------------------------------------------ |--------|--------------:|---------:|---------:|------:|------:|------:|----------:|
|                                     GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsNull | PR     |      52.54 ns | 0.162 ns | 0.135 ns |     - |     - |     - |         - |
|                                     GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsNull | master |      80.97 ns | 0.409 ns | 0.382 ns |     - |     - |     - |         - |
|    GetFireTimeAfter_RepeatCountIsGreaterThanZeroAndLessThanNumberOfTimesExecuted_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | PR     |      35.04 ns | 0.049 ns | 0.043 ns |     - |     - |     - |         - |
|    GetFireTimeAfter_RepeatCountIsGreaterThanZeroAndLessThanNumberOfTimesExecuted_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | master |      42.48 ns | 0.073 ns | 0.061 ns |     - |     - |     - |         - |
|                                                            GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsNull | PR     |     126.28 ns | 0.320 ns | 0.267 ns |     - |     - |     - |         - |
|                                                            GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsNull | master |     148.02 ns | 0.197 ns | 0.154 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | PR     |      67.53 ns | 0.118 ns | 0.110 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMaxValue | master |      74.28 ns | 0.099 ns | 0.077 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMinValue | PR     |      24.76 ns | 0.063 ns | 0.055 ns |     - |     - |     - |         - |
|                                 GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndAfterStartTimeUtc_EndTimeUtcIsMinValue | master |      26.37 ns | 0.046 ns | 0.038 ns |     - |     - |     - |         - |
|                                                                    GetFireTimeAfter_RepeatCountIsZero_AfterTimeUtcIsNotNullAndAfterStartTimeUtc | PR     |      19.01 ns | 0.035 ns | 0.027 ns |     - |     - |     - |         - |
|                                                                    GetFireTimeAfter_RepeatCountIsZero_AfterTimeUtcIsNotNullAndAfterStartTimeUtc | master |      18.98 ns | 0.042 ns | 0.039 ns |     - |     - |     - |         - |
|                                                        GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsMaxValue | PR     |     142.08 ns | 0.253 ns | 0.225 ns |     - |     - |     - |         - |
|                                                        GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNull_EndTimeUtcIsMaxValue | master |     149.71 ns | 0.257 ns | 0.200 ns |     - |     - |     - |         - |
|                                    GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndBeforeStartTimeUtc_EndTimeUtcIsNull | PR     |      25.75 ns | 0.125 ns | 0.111 ns |     - |     - |     - |         - |
|                                    GetFireTimeAfter_RepeatCountIsRepeatIndefinitely_AfterTimeUtcIsNotNullAndBeforeStartTimeUtc_EndTimeUtcIsNull | master |      30.93 ns | 0.042 ns | 0.035 ns |     - |     - |     - |         - |
|                                                                       Triggered_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsNull_CalendarIsNull | PR     |      57.06 ns | 0.188 ns | 0.157 ns |     - |     - |     - |         - |
|                                                                       Triggered_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsNull_CalendarIsNull | master |      77.78 ns | 0.408 ns | 0.362 ns |     - |     - |     - |         - |
|                                                                                                UpdateAfterMisfire_SmartPolicy_RepeatCountIsZero | PR     |      85.21 ns | 0.287 ns | 0.255 ns |     - |     - |     - |         - |
|                                                                                                UpdateAfterMisfire_SmartPolicy_RepeatCountIsZero | master |      86.49 ns | 0.087 ns | 0.068 ns |     - |     - |     - |         - |
|                                                                                  UpdateAfterMisfire_SmartPolicy_RepeatCountIsRepeatIndefinitely | PR     |     156.71 ns | 0.670 ns | 0.627 ns |     - |     - |     - |         - |
|                                                                                  UpdateAfterMisfire_SmartPolicy_RepeatCountIsRepeatIndefinitely | master |     179.06 ns | 0.332 ns | 0.294 ns |     - |     - |     - |         - |
|                                                                                     UpdateAfterMisfire_SmartPolicy_RepeatCountIsGreaterThanZero | PR     |     123.61 ns | 0.121 ns | 0.101 ns |     - |     - |     - |         - |
|                                                                                     UpdateAfterMisfire_SmartPolicy_RepeatCountIsGreaterThanZero | master |     123.84 ns | 0.584 ns | 0.517 ns |     - |     - |     - |         - |
|                                                                 UpdateAfterMisfire_FireNow_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMinValue | PR     |     104.00 ns | 0.192 ns | 0.160 ns |     - |     - |     - |         - |
|                                                                 UpdateAfterMisfire_FireNow_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMinValue | master |     120.33 ns | 0.378 ns | 0.316 ns |     - |     - |     - |         - |
|                                                                           UpdateAfterMisfire_FireNow_RepeatCountIsMaxValue_EndTimeUtcIsMaxValue | PR     |     138.39 ns | 0.230 ns | 0.204 ns |     - |     - |     - |         - |
|                                                                           UpdateAfterMisfire_FireNow_RepeatCountIsMaxValue_EndTimeUtcIsMaxValue | master |     139.58 ns | 0.297 ns | 0.278 ns |     - |     - |     - |         - |
|                                         UpdateAfterMisfire_RescheduleNextWithExistingCount_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMaxValue | PR     |     146.03 ns | 0.924 ns | 0.864 ns |     - |     - |     - |         - |
|                                         UpdateAfterMisfire_RescheduleNextWithExistingCount_RepeatCountIsRepeatIndefinitely_EndTimeUtcIsMaxValue | master |     160.62 ns | 0.255 ns | 0.226 ns |     - |     - |     - |         - |
</details>